implement add-todo functionality
