+++ src/index.html
@@ -22,8 +22,9 @@
         <p>Todays To do</p>
         <i class="fa-solid fa-arrows-rotate refresh-icon"></i>
       </header>
-      <form>
-        <input type="text" placeholder="Add to your list" />
+      <form class="form">
+        <input class="input" type="text" placeholder="Add to your list" />
+        <button class="submit__button" type="submit">Add Todo</button>
       </form>
       <div class="todoList__wrapper">
       </div>
